+++ common/src/main/java/com/faboslav/structurify/common/config/client/gui/StructureSetsConfigScreen.java
@@ -217,13 +217,10 @@ public final class StructureSetsConfigScreen
 
 				this.structureSetOptions.put(structureSetStringId, new AbstractMap.SimpleEntry<>(overrideGlobalSpacingAndSeparationModifierOption, spacingAndSeparationOption));
 			}
-
-			//currentGroupBuilder.collapsed(true);
 		}
 
 		if (currentGroupBuilder != null) {
 			OptionGroup buildedGroup = currentGroupBuilder.build();
-			//currentGroupBuilder.collapsed(true);
 			this.optionGroups.put(buildedGroup.name().getString(), buildedGroup);
 			optionGroups.add(buildedGroup);
 		}