+++ common/src/main/java/com/faboslav/structurify/common/config/client/gui/StructurifyConfigScreenState.java
@@ -1,8 +1,10 @@
 package com.faboslav.structurify.common.config.client.gui;
 
+import java.util.Map;
+
 public record StructurifyConfigScreenState(
 	String lastSearchText,
-	double lastScrollAmount
-)
-{
+	double lastScrollAmount,
+	Map<String, Boolean> collapsedGroups
+) {
 }